chore: move token utility functions to bench
