+++ governance/test-sdk/Cargo.toml
@@ -18,6 +18,7 @@ solana-program = "1.7.7"
 thiserror = "1.0"
 solana-program-test = "1.7.7"
 solana-sdk = "1.7.7"
+spl-token = { version = "3.2", path = "../../token/program", features = [ "no-entrypoint" ] }
 
 [lib]
 crate-type = ["cdylib", "lib"]
