+++ governance/test-sdk/src/cookies.rs
@@ -1,0 +1,6 @@
+use solana_program::pubkey::Pubkey;
+
+#[derive(Debug)]
+pub struct TokenAccountCookie {
+    pub address: Pubkey,
+}
