+++ governance/test-sdk/src/lib.rs
@@ -1,5 +1,7 @@
+use cookies::TokenAccountCookie;
 use solana_program::{
-    instruction::Instruction, program_error::ProgramError, pubkey::Pubkey, rent::Rent,
+    instruction::Instruction, program_error::ProgramError, program_pack::Pack, pubkey::Pubkey,
+    rent::Rent, system_instruction,
 };
 use solana_program_test::{ProgramTest, ProgramTestContext};
 use solana_sdk::{
@@ -10,6 +12,7 @@ use tools::clone_keypair;
 
 use crate::tools::map_transaction_error;
 
+pub mod cookies;
 pub mod tools;
 
 /// Program's test bench which captures test context, rent and payer and common utility functions
@@ -80,5 +83,111 @@ impl ProgramTestBench {
             .map_err(map_transaction_error)?;
 
         Ok(())
+    }
+
+    pub async fn create_mint(&mut self, mint_keypair: &Keypair, mint_authority: &Pubkey) {
+        let mint_rent = self.rent.minimum_balance(spl_token::state::Mint::LEN);
+
+        let instructions = [
+            system_instruction::create_account(
+                &self.context.payer.pubkey(),
+                &mint_keypair.pubkey(),
+                mint_rent,
+                spl_token::state::Mint::LEN as u64,
+                &spl_token::id(),
+            ),
+            spl_token::instruction::initialize_mint(
+                &spl_token::id(),
+                &mint_keypair.pubkey(),
+                mint_authority,
+                None,
+                0,
+            )
+            .unwrap(),
+        ];
+
+        self.process_transaction(&instructions, Some(&[mint_keypair]))
+            .await
+            .unwrap();
+    }
+
+    #[allow(dead_code)]
+    pub async fn create_empty_token_account(
+        &mut self,
+        token_account_keypair: &Keypair,
+        token_mint: &Pubkey,
+        owner: &Pubkey,
+    ) {
+        let create_account_instruction = system_instruction::create_account(
+            &self.context.payer.pubkey(),
+            &token_account_keypair.pubkey(),
+            self.rent
+                .minimum_balance(spl_token::state::Account::get_packed_len()),
+            spl_token::state::Account::get_packed_len() as u64,
+            &spl_token::id(),
+        );
+
+        let initialize_account_instruction = spl_token::instruction::initialize_account(
+            &spl_token::id(),
+            &token_account_keypair.pubkey(),
+            token_mint,
+            owner,
+        )
+        .unwrap();
+
+        self.process_transaction(
+            &[create_account_instruction, initialize_account_instruction],
+            Some(&[token_account_keypair]),
+        )
+        .await
+        .unwrap();
+    }
+
+    #[allow(dead_code)]
+    pub async fn with_token_account(
+        &mut self,
+        token_mint: &Pubkey,
+        owner: &Pubkey,
+        token_mint_authority: &Keypair,
+        amount: u64,
+    ) -> TokenAccountCookie {
+        let token_account_keypair = Keypair::new();
+
+        self.create_empty_token_account(&token_account_keypair, token_mint, owner)
+            .await;
+
+        self.mint_tokens(
+            token_mint,
+            token_mint_authority,
+            &token_account_keypair.pubkey(),
+            amount,
+        )
+        .await;
+
+        return TokenAccountCookie {
+            address: token_account_keypair.pubkey(),
+        };
+    }
+
+    pub async fn mint_tokens(
+        &mut self,
+        token_mint: &Pubkey,
+        token_mint_authority: &Keypair,
+        token_account: &Pubkey,
+        amount: u64,
+    ) {
+        let mint_instruction = spl_token::instruction::mint_to(
+            &spl_token::id(),
+            token_mint,
+            token_account,
+            &token_mint_authority.pubkey(),
+            &[],
+            amount,
+        )
+        .unwrap();
+
+        self.process_transaction(&[mint_instruction], Some(&[token_mint_authority]))
+            .await
+            .unwrap();
     }
 }
